requirements.txt
psycopg[binary]>=3.2,<3.3
gspread>=6.1,<7
google-auth>=2.39,<3
